+++ documentation/docs/pages/developers/smolmix/_meta.json
@@ -1,0 +1,7 @@
+{
+  "getting-started": "Getting Started",
+  "dns": "DNS Resolution",
+  "tls": "TLS",
+  "http": "HTTP Client",
+  "advanced": "Advanced Patterns"
+}
