+++ documentation/docs/pages/developers/smolmix/advanced/_meta.json
@@ -1,0 +1,4 @@
+{
+  "websocket": "WebSocket",
+  "libp2p": "libp2p Transport"
+}
