+++ documentation/docs/pages/developers/smolmix/advanced/websocket.mdx
@@ -1,0 +1,105 @@
+---
+title: "WebSocket Over the Mixnet"
+description: "Connect to WebSocket servers through the Nym mixnet using tokio-tungstenite and smolmix-tls. A copy-paste pattern, not a separate crate."
+schemaType: "TechArticle"
+section: "Developers"
+lastUpdated: "2026-04-28"
+---
+
+# WebSocket
+
+import { Callout } from 'nextra/components'
+import { VersionBanner } from '../../../../components/version-banner'
+
+<VersionBanner />
+
+WebSocket over smolmix doesn't require a dedicated crate — it's just function composition. `tokio-tungstenite`'s `client_async()` accepts any `AsyncRead + AsyncWrite` stream, and `TlsStream<TcpStream>` from smolmix-tls qualifies.
+
+## The pattern
+
+```text
+smolmix_dns::resolve(host, port)    → DNS through tunnel
+tunnel.tcp_connect(addr)            → TCP through mixnet
+smolmix_tls::connect(tcp, host)     → TLS handshake
+client_async(request, tls_stream)   → WebSocket upgrade
+```
+
+Four function calls, no new types, no trait implementations.
+
+## Dependencies
+
+```toml
+[dependencies]
+smolmix = "0.0.1"
+smolmix-dns = "0.0.1"
+smolmix-tls = "0.0.1"
+tokio-tungstenite = "0.24"
+futures = "0.3"
+```
+
+## Full example
+
+```rust
+use futures::{SinkExt, StreamExt};
+use smolmix::Tunnel;
+use tokio_tungstenite::tungstenite::Message;
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    smolmix::init_logging();
+    rustls::crypto::ring::default_provider()
+        .install_default()
+        .expect("Failed to install rustls crypto provider");
+
+    let tunnel = Tunnel::new().await?;
+
+    let host = "ws.postman-echo.com";
+    let port = 443;
+
+    // 1. DNS through the tunnel
+    let addrs = smolmix_dns::resolve(&tunnel, host, port).await?;
+    let addr = addrs[0];
+
+    // 2. TCP through the mixnet
+    let tcp = tunnel.tcp_connect(addr).await?;
+
+    // 3. TLS handshake
+    let tls = smolmix_tls::connect(tcp, host).await?;
+
+    // 4. WebSocket upgrade
+    let request = format!("wss://{host}/raw");
+    let (mut ws, _response) = tokio_tungstenite::client_async(request, tls).await?;
+
+    // Send and receive
+    ws.send(Message::Text("Hello from the Nym mixnet!".into())).await?;
+    let reply = ws.next().await.ok_or("no reply")??;
+    println!("Echo: {}", reply.into_text()?);
+
+    ws.close(None).await?;
+    tunnel.shutdown().await;
+    Ok(())
+}
+```
+
+## Why not a crate?
+
+The entire value of a `smolmix-tungstenite` crate would be wrapping these 4 calls in a `connect()` function — saving ~10 lines. That's not enough to justify a published crate with a semver contract, version maintenance, and documentation burden.
+
+If tokio-tungstenite changes its API, you update your 10 lines. If it were a crate, you'd wait for a release.
+
+<Callout type="info">
+A complete runnable example lives at [`smolmix/core/examples/websocket.rs`](https://github.com/nymtech/nym/blob/develop/smolmix/core/examples/websocket.rs) with clearnet-vs-mixnet timing comparison.
+</Callout>
+
+## Plain WebSocket (`ws://`)
+
+For unencrypted WebSocket, skip the TLS step:
+
+```rust
+let tcp = tunnel.tcp_connect(addr).await?;
+let (ws, _) = tokio_tungstenite::client_async("ws://echo.example.com/ws", tcp).await?;
+```
+
+<Callout type="warning">
+Without TLS, the IPR exit gateway can read your WebSocket traffic. Only use `ws://` for non-sensitive data or when the application layer provides its own encryption.
+</Callout>
