+++ documentation/docs/pages/developers/smolmix/getting-started.mdx
@@ -1,0 +1,114 @@
+---
+title: "Getting Started with smolmix"
+description: "Create your first smolmix Tunnel, make a TCP connection through the Nym mixnet, and verify your traffic exits from the IPR gateway."
+schemaType: "HowTo"
+section: "Developers"
+lastUpdated: "2026-04-28"
+---
+
+# Getting Started
+
+import { Callout } from 'nextra/components'
+import { VersionBanner } from '../../../components/version-banner'
+import { RUST_MSRV } from '../../../components/versions'
+
+<VersionBanner />
+
+This guide walks you through creating a smolmix `Tunnel` and making your first TCP connection through the Nym mixnet.
+
+## Prerequisites
+
+- Rust toolchain ({RUST_MSRV}+)
+- A working internet connection (the tunnel connects to the live Nym mixnet)
+
+## Step 1: Set up the project
+
+```sh
+cargo init smolmix-hello
+cd smolmix-hello
+```
+
+Add dependencies to `Cargo.toml`:
+
+```toml
+[dependencies]
+smolmix = "0.0.1"
+nym-bin-common = { version = "1.20.4", features = ["basic_tracing"] }
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "io-util"] }
+blake3 = "=1.7.0"
+```
+
+<Callout type="warning">
+**Temporary pin required.** You must pin `blake3 = "=1.7.0"` to avoid a build failure caused by a transitive `digest` version conflict. This will be resolved in a future release.
+</Callout>
+
+## Step 2: Connect and make a request
+
+```rust
+use smolmix::Tunnel;
+use tokio::io::{AsyncReadExt, AsyncWriteExt};
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    // Set up logging so you can see connection progress
+    smolmix::init_logging();
+
+    // Create a tunnel — this connects to the Nym mixnet and allocates
+    // an IP address from the IPR exit gateway.
+    let tunnel = Tunnel::new().await?;
+    println!("Tunnel ready, allocated IP: {:?}", tunnel.allocated_ips());
+
+    // Open a TCP connection to Cloudflare's diagnostic endpoint.
+    // The connection exits at the IPR gateway, not from your machine.
+    let mut tcp = tunnel.tcp_connect("1.1.1.1:80".parse()?).await?;
+
+    // Send a raw HTTP/1.1 request
+    tcp.write_all(
+        b"GET /cdn-cgi/trace HTTP/1.1\r\nHost: 1.1.1.1\r\nConnection: close\r\n\r\n"
+    ).await?;
+
+    // Read the response
+    let mut buf = Vec::new();
+    tcp.read_to_end(&mut buf).await?;
+    let response = String::from_utf8_lossy(&buf);
+    println!("{response}");
+
+    // The `ip=` line shows the exit gateway's IP, not yours.
+    tunnel.shutdown().await;
+    Ok(())
+}
+```
+
+## Step 3: Run it
+
+```sh
+cargo run
+```
+
+You'll see logging output as the mixnet client connects, followed by the HTTP response. The `ip=` field shows the IPR gateway's IP address — your real IP is hidden.
+
+## What's happening under the hood
+
+1. `Tunnel::new()` connects a Nym `MixnetClient`, discovers an IPR exit gateway, and sets up a userspace TCP/IP stack (smoltcp)
+2. `tcp_connect()` creates a TCP socket managed by the smoltcp reactor — the TCP handshake (SYN/ACK) travels through the mixnet
+3. All bytes written to the stream are wrapped as IP packets, Sphinx-encrypted, mixed through 5 nodes, and reassembled at the exit gateway
+4. The exit gateway forwards the TCP connection to `1.1.1.1:80` as normal internet traffic
+
+## Next steps
+
+- [DNS Resolution](./dns) — resolve hostnames through the tunnel (avoid DNS leaks)
+- [TLS](./tls) — encrypt the final hop between IPR and destination
+- [HTTP Client](./http) — high-level HTTP requests via hyper
+
+## Targeting a specific exit gateway
+
+Pass `--ipr <ADDRESS>` to target a specific IPR node:
+
+```rust
+use smolmix::{Tunnel, Recipient};
+
+let ipr: Recipient = "gateway-address...".parse()?;
+let tunnel = Tunnel::new_with_ipr(ipr).await?;
+```
+
+This is useful for testing or when you want a specific exit location.
