+++ documentation/docs/pages/developers/smolmix/tls.mdx
@@ -1,0 +1,101 @@
+---
+title: "smolmix-tls: TLS Over the Mixnet"
+description: "Encrypt the final hop between the IPR exit gateway and your destination using smolmix-tls. Pre-configured TLS with webpki root certificates."
+schemaType: "TechArticle"
+section: "Developers"
+lastUpdated: "2026-04-28"
+---
+
+# TLS
+
+import { Callout } from 'nextra/components'
+import { VersionBanner } from '../../../components/version-banner'
+
+<VersionBanner />
+
+Traffic between the IPR exit gateway and the remote host travels as normal internet traffic. Without TLS, the exit gateway (and anyone between it and the destination) can read your data. `smolmix-tls` provides pre-configured TLS that works directly over `smolmix::TcpStream`.
+
+<Callout type="warning">
+**Always use TLS for sensitive traffic.** The mixnet protects your identity (who is connecting), but without TLS the exit gateway can see the payload content (what you're sending).
+</Callout>
+
+## Installation
+
+```toml
+[dependencies]
+smolmix = "0.0.1"
+smolmix-tls = "0.0.1"
+```
+
+## Basic usage
+
+```rust
+use smolmix::Tunnel;
+use tokio::io::{AsyncReadExt, AsyncWriteExt};
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let tunnel = Tunnel::new().await?;
+
+    // Connect TCP through the mixnet
+    let tcp = tunnel.tcp_connect("93.184.216.34:443".parse()?).await?;
+
+    // TLS handshake — SNI set to "example.com"
+    let mut tls = smolmix_tls::connect(tcp, "example.com").await?;
+
+    // Now the connection is encrypted end-to-end
+    tls.write_all(b"GET / HTTP/1.1\r\nHost: example.com\r\nConnection: close\r\n\r\n").await?;
+    let mut buf = Vec::new();
+    tls.read_to_end(&mut buf).await?;
+    println!("{}", String::from_utf8_lossy(&buf));
+
+    tunnel.shutdown().await;
+    Ok(())
+}
+```
+
+## Reusing the connector
+
+`smolmix_tls::connect()` rebuilds the root certificate store on each call. For multiple connections, create a connector once and reuse it:
+
+```rust
+// Create once — wraps an Arc<ClientConfig>, cheap to clone
+let tls_connector = smolmix_tls::connector();
+
+// Use for multiple connections
+let tcp1 = tunnel.tcp_connect(addr1).await?;
+let stream1 = smolmix_tls::connect_with(&tls_connector, tcp1, "a.com").await?;
+
+let tcp2 = tunnel.tcp_connect(addr2).await?;
+let stream2 = smolmix_tls::connect_with(&tls_connector, tcp2, "b.com").await?;
+```
+
+<Callout type="info">
+`smolmix-hyper` handles this automatically — its `SmolmixConnector` caches a `TlsConnector` internally.
+</Callout>
+
+## How it works
+
+`tokio-rustls` works with anything implementing tokio's `AsyncRead + AsyncWrite`. Since `smolmix::TcpStream` (from tokio-smoltcp) implements these traits, no adapter is needed — this crate is pure configuration:
+
+1. Build a `ClientConfig` with Mozilla's root CA bundle ([webpki-roots](https://docs.rs/webpki-roots))
+2. Wrap in a `TlsConnector`
+3. Call `.connect(server_name, tcp)` to perform the handshake
+
+No certificate verification bypass is possible — this is intentional for a privacy tool.
+
+## API reference
+
+| Function | Description |
+|---|---|
+| `connector()` | Build a `TlsConnector` with webpki roots (reusable) |
+| `connect(tcp, hostname)` | One-shot TLS handshake (fresh connector) |
+| `connect_with(&tls, tcp, hostname)` | TLS handshake with existing connector |
+| `TlsStream` (re-export) | The encrypted stream type |
+| `TlsConnector` (re-export) | The connector type (clone-cheap) |
+
+## When to use this directly
+
+- **Use `smolmix-tls` directly** when you're building custom protocols over TCP and need manual TLS control
+- **Use `smolmix-hyper`** for HTTP/HTTPS — it handles DNS + TLS + HTTP together
+- **Use tokio-tungstenite + `smolmix-tls`** for WebSocket — see [Advanced: WebSocket](./advanced/websocket)
