+++ smolmix/hyper/src/lib.rs
@@ -1,15 +1,40 @@
-// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// Copyright 2024-2026 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-2.0-only
 
 //! HTTP client routing all traffic through the Nym mixnet.
 //!
-//! This crate wraps [hyper-util]'s `Client` with a newtype [`Client`] that
-//! routes DNS, TCP, and TLS through a smolmix [`Tunnel`]. It feels like a
-//! drop-in replacement for `hyper_util::client::legacy::Client`.
+//! # What this crate does
 //!
-//! # Quick start
+//! Wraps [hyper-util]'s legacy `Client` with a [`SmolmixConnector`] that
+//! routes DNS resolution, TCP connections, and TLS handshakes through a
+//! smolmix [`Tunnel`]. From the outside it behaves like a normal HTTP client,
+//! but all traffic travels through the mixnet.
 //!
-//! ```ignore
+//! # How it works
+//!
+//! hyper-util's `Client` uses the [`tower::Service<Uri>`] trait to open
+//! connections. [`SmolmixConnector`] implements this: given a URI, it resolves
+//! the hostname via [`smolmix_dns`], connects TCP via the tunnel, and wraps
+//! in TLS for `https://` URIs:
+//!
+//! ```text
+//! SmolmixConnector::call(uri)
+//!   → smolmix_dns::resolve(host, port)    DNS through tunnel
+//!   → tunnel.tcp_connect(addr)            TCP through mixnet
+//!   → smolmix_tls::connect_with(tls, tcp, host)   TLS if https
+//!   → MaybeTlsStream::Plain { TcpStream }
+//!     or MaybeTlsStream::Tls { TlsStream<TcpStream> }
+//!   → TokioIo<MaybeTlsStream>   (implements hyper's Read/Write/Connection)
+//! ```
+//!
+//! [`MaybeTlsStream`] is a two-variant enum with [`pin_project_lite`] for safe
+//! pin projection through `AsyncRead`/`AsyncWrite`. hyper-util's `TokioIo`
+//! wraps it to satisfy hyper's own I/O traits.
+//!
+//! # Quick start (GET)
+//!
+//! ```no_run
+//! # async fn example() -> Result<(), Box<dyn std::error::Error>> {
 //! use smolmix_hyper::{Client, Request, EmptyBody, BodyExt};
 //! use bytes::Bytes;
 //!
@@ -21,10 +46,53 @@
 //!     .body(EmptyBody::<Bytes>::new())?;
 //! let resp = client.request(req).await?;
 //! let body = resp.into_body().collect().await?.to_bytes();
+//! println!("{}", String::from_utf8_lossy(&body));
+//! # Ok(())
+//! # }
 //! ```
+//!
+//! # Sending request bodies (POST, PUT, etc.)
+//!
+//! The convenience [`Client`] wrapper uses `Empty<Bytes>` as its body type,
+//! which is suitable for GET/HEAD/DELETE. For requests that carry a body,
+//! construct a hyper-util client directly with [`SmolmixConnector`]:
+//!
+//! ```no_run
+//! # async fn example() -> Result<(), Box<dyn std::error::Error>> {
+//! use http_body_util::Full;
+//! use hyper_util::{client::legacy, rt::TokioExecutor};
+//! use bytes::Bytes;
+//! use smolmix_hyper::SmolmixConnector;
+//!
+//! let tunnel = smolmix::Tunnel::new().await?;
+//! let connector = SmolmixConnector::new(&tunnel);
+//! let client = legacy::Client::builder(TokioExecutor::new())
+//!     .build::<_, Full<Bytes>>(connector);
+//!
+//! let body = Full::new(Bytes::from(r#"{"key": "value"}"#));
+//! let req = hyper::Request::post("https://httpbin.org/post")
+//!     .header("Host", "httpbin.org")
+//!     .header("Content-Type", "application/json")
+//!     .body(body)?;
+//! let resp = client.request(req).await?;
+//! # Ok(())
+//! # }
+//! ```
+//!
+//! # Re-exports
+//!
+//! This crate re-exports the most commonly needed types so you don't need
+//! `hyper`, `http-body-util`, or `bytes` in your `Cargo.toml` for basic use:
+//!
+//! - [`Request`], [`Response`], [`StatusCode`], [`Uri`] — from hyper
+//! - [`BodyExt`], [`EmptyBody`] — from http-body-util
+//! - [`bytes`] — the bytes crate
+//!
+//! [hyper-util]: https://docs.rs/hyper-util
+//! [`pin_project_lite`]: https://docs.rs/pin-project-lite
 
 mod connector;
-pub mod tls_stream;
+mod tls_stream;
 
 use std::ops::Deref;
 
@@ -35,12 +103,25 @@ use hyper_util::rt::TokioExecutor;
 
 use smolmix::Tunnel;
 
-// Re-exports so users don't need hyper/http-body-util/bytes in their Cargo.toml
+/// Re-exported [`bytes`](https://docs.rs/bytes) crate for constructing request bodies.
 pub use bytes;
-pub use http_body_util::{BodyExt, Empty as EmptyBody};
+
+/// Extension trait for consuming HTTP response bodies. Provides `.collect()`,
+/// `.frame()`, etc.
+pub use http_body_util::BodyExt;
+
+/// An empty HTTP body. Use `EmptyBody::<Bytes>::new()` for GET/HEAD requests.
+pub use http_body_util::Empty as EmptyBody;
+
+/// Re-exported hyper types for building requests without depending on hyper directly.
 pub use hyper::{Request, Response, StatusCode, Uri};
 
+/// The connector implementing `tower::Service<Uri>`. Use this directly when you
+/// need a custom body type (e.g. `Full<Bytes>` for POST requests).
 pub use connector::SmolmixConnector;
+
+/// A stream that may or may not be wrapped in TLS. Returned by [`SmolmixConnector`]
+/// — you won't normally interact with this directly.
 pub use tls_stream::MaybeTlsStream;
 
 /// Inner hyper-util client type alias for readability.
