--- smolmix/tungstenite/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "smolmix-tungstenite"
-version = "0.0.1"
-edition = "2021"
-license.workspace = true
-
-[dependencies]
-smolmix = { workspace = true }
-smolmix-dns = { workspace = true }
-smolmix-tls = { workspace = true }
-tokio-smoltcp = "0.5"
-tokio-tungstenite = { workspace = true }
-
-[dev-dependencies]
-futures = { workspace = true }
-rustls = { workspace = true }
-tokio = { workspace = true, features = ["macros", "rt-multi-thread", "time", "net"] }
-tracing = { workspace = true }
